# [tensor][perf] Replace list comprehension with *.