@tinyb0t 2023-07-03
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         2               2
------------------------------------------------------------
lines added in the tinygrad folder: 0
```