@geohot 2023-07-03
neat. can't believe I wrote `[1 for _` lol