@ttsugriy 2023-07-03
> neat. can't believe I wrote `[1 for _` lol

it happens to the best of us, although I suspect that you've done it intentionally to give bragging right opportunities for folks like me :)